remove_all test code compile warning
Former-commit-id: 5060cfbc14547439a083e86264f7db861523bc2a

+++ src/test/scriptdbex_tests.cpp
@@ -84,7 +84,7 @@ void CScriptDBTest::CheckRecordCount(CScriptDBViewCache* pViewCache,size_t nComp
 	BOOST_CHECK(pViewCache);
 	int nCount = 0;
 	BOOST_CHECK(pViewCache->GetScriptCount(nCount));
-	BOOST_CHECK(nCount == nComparCount);
+	BOOST_CHECK((unsigned int)nCount == nComparCount);
 }
 
 void CScriptDBTest::Flush(CScriptDBViewCache* pViewCache1, CScriptDBViewCache* pViewCache2,
